requirements.txt
pandas==1.5.3
azure-common==1.1.28
azure-mgmt-keyvault==10.2.0
azure-mgmt-resource==22.0.0
azure-mgmt-subscription==3.1.1
azure-storage-blob==12.14.1
azure-identity==1.12.0
azure-mgmt-storage==21.0.0
msrestazure==0.6.4
azure-keyvault==4.2.0
azure-mgmt-automation==1.0.0
azure-mgmt-advisor==9.0.0

azure.identity
azure.mgmt.compute
azure.mgmt.network
azure.mgmt.storage
azure.mgmt.keyvault
azure.mgmt.subscription
azure.mgmt.resource
azure.mgmt.rdbms
azure.mgmt.sql
azure.mgmt.frontdoor
azure.mgmt.web
azure.mgmt.trafficmanager
azure.mgmt.cosmosdb
azure.mgmt.dns
azure.mgmt.privatedns
azure.mgmt.apimanagement
azure.mgmt.servicebus
azure.mgmt.databricks
azure.mgmt.msi
pandas
